py-tavern-api/requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorlog==4.1.0
docopt==0.6.2
idna==2.9
jmespath==0.9.5
more-itertools==8.2.0
packaging==20.3
paho-mqtt==1.3.1
pbr==5.4.5
pluggy==0.13.1
py==1.8.1
PyJWT==1.7.1
pykwalify==1.7.0
pyparsing==2.4.7
pytest==5.4.1
python-box==4.2.2
python-dateutil==2.8.1
python-dotenv==0.13.0
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
stevedore==1.32.0
tavern==1.0.0
toml==0.10.0
urllib3==1.25.8
wcwidth==0.1.9
